B-tree: Make space for values configurable (#2814)
* initial working version

* make to and from long more clear

* make it working for negative keys and values

* nextLong(bounds) only for more recent JDKs

* fix javadoc

* Use long for pillar node id stack variables

* use long where it could be a pillar or tower node

* simulate large amount of pillar nodes

* remove second check for pillar nodes

* use 0 as start again

* minor

* make empty value configurable

* changes for review

---------

+++ core/src/main/java/com/graphhopper/reader/osm/SegmentNode.java
@@ -22,10 +22,10 @@ import java.util.Map;
 
 class SegmentNode {
     long osmNodeId;
-    int id;
+    long id;
     Map<String, Object> tags;
 
-    public SegmentNode(long osmNodeId, int id, Map<String, Object> tags) {
+    public SegmentNode(long osmNodeId, long id, Map<String, Object> tags) {
         this.osmNodeId = osmNodeId;
         this.id = id;
         this.tags = tags;
